Font provider can be accessed by old FontDatabase interface
The system now allows a single entry point for loading and using fonts.

Tha cmake option ROCKET_WITH_FREETYPE allows to build with our without freetype lib.

Samples have been updated to use this single entry point as well

+++ Source/Core/BitmapFont/FontFaceHandle.cpp
@@ -297,7 +297,7 @@ void FontFaceHandle::GenerateMetrics(BM_Font *bm_face)
 	/*underline_thickness = FT_MulFix(ft_face->underline_thickness, ft_face->size->metrics.y_scale) / float(1 << 6);
 	underline_thickness = Math::Max(underline_thickness, 1.0f);
 */
-	baseline += underline_position / 1.5f;
+	baseline += int( underline_position / 1.5f );
 	underline_thickness = 1.0f;
 	average_advance = 0;
 	for (FontGlyphMap::iterator i = glyphs.begin(); i != glyphs.end(); ++i)
